refactor(search): rename date utils file to date.ts
