+++ src/utils/index.ts
@@ -2,4 +2,4 @@ export * from './ui';
 export * from './validation';
 export * from './languages';
 export * from './getQuery';
-export * from './stringToDateFormat';
+export * from './date';
